+++ extensions-leaflet-map.php
@@ -4,7 +4,7 @@
  * Description:       Extends the WordPress Plugin <a href="https://wordpress.org/plugins/leaflet-map/">Leaflet Map</a> with Leaflet Plugins and other functions.
  * Plugin URI:        https://leafext.de/en/
  * Update URI:        https://github.com/hupe13/extensions-leaflet-map-github
- * Version:           4.12-260214
+ * Version:           4.12-260218
  * Requires PHP:      8.1
  * Requires Plugins:  leaflet-map
  * Author:            hupe13